requirements.txt
streamlit
pandas
pypdf
scikit-learn
streamlit-mic-recorder
google-generativeai
numpy